The inetrupt drive version works however polling verion seem unreliable. Clock seem only to be being held down for very shortperiod by the KB.

+++ stm8/stm8s_it.c
@@ -29,7 +29,7 @@
 /* Private function prototypes -----------------------------------------------*/
 
 
-extern __IO uint32_t TIM4_Clock;
+extern __IO unsigned long TIM4_Clock;
 void ClkIntHandler(void);
 
 extern	void  	GPIO_D_Handler(void);
